move featureflag to its own source
